Some refactoring, improved logging, and add cancellation.
